Extract connection related logic into abstract ConnectionFetcher class
